# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/caregivers": {
      "post": {
        "description": "Crea el perfil profesional en estado pending. No visible en el marketplace hasta UC-19.",
        "operationId": "CaregiverController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCaregiverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-02 · Registrar cuidador",
        "tags": [
          "Caregivers"
        ]
      }
    },
    "/api/v1/caregivers/me": {
      "put": {
        "description": "Corrige los datos y re-envía. Solo desde estado rejected: el perfil vuelve a pending, se limpia el motivo de rechazo y las certificaciones vuelven a \"no verificada\". Naturalmente idempotente.",
        "operationId": "CaregiverController_resubmit_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCaregiverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-02 A2 · Re-enviar postulación tras rechazo",
        "tags": [
          "Caregivers"
        ]
      },
      "get": {
        "operationId": "CaregiverController_myProfile_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-02 · Ver mi perfil y estado de aprobación",
        "tags": [
          "Caregivers"
        ]
      }
    },
...
    "/api/v1/marketplace/caregivers": {
      "get": {
        "operationId": "MarketplaceController_search_v1",
        "parameters": [
          {
            "name": "careType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "elder-care",
                "post-surgical",
                "chronic-illness",
                "disability",
                "palliative",
                "pediatric",
                "rehabilitation",
                "companionship"
              ]
            }
          },
          {
            "name": "modality",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "home",
                "hospital"
              ]
            }
          },
          {
            "name": "zone",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Palermo",
              "type": "string"
            }
          },
          {
            "name": "minRatePerHour",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1000,
              "type": "number"
            }
          },
          {
            "name": "maxRatePerHour",
            "required": false,
            "in": "query",
            "schema": {
              "example": 5000,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaregiverCardDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-06 · Buscar cuidadores (solo aprobados, filtros combinables)",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/api/v1/marketplace/caregivers/{id}": {
      "get": {
        "operationId": "MarketplaceController_profile_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-07 · Ver perfil de cuidador",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/api/v1/favorites": {
      "get": {
        "operationId": "MarketplaceController_favorites_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaregiverCardDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-08 · Listar favoritos",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/api/v1/favorites/{caregiverId}": {
      "post": {
        "operationId": "MarketplaceController_addFavorite_v1",
        "parameters": [
          {
            "name": "caregiverId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-08 · Marcar favorito (idempotente)",
        "tags": [
          "Marketplace"
        ]
      },
      "delete": {
        "operationId": "MarketplaceController_removeFavorite_v1",
        "parameters": [
          {
            "name": "caregiverId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-08 · Quitar favorito",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/api/v1/hiring-requests": {
      "post": {
        "operationId": "MarketplaceController_createRequest_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-09 · Crear solicitud de contratación (una por paciente)",
        "tags": [
          "Marketplace"
        ]
      },
      "get": {
        "operationId": "MarketplaceController_myRequests_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RequestResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-09 · Mis solicitudes",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/api/v1/hiring-requests/{id}/cancel": {
      "post": {
        "description": "Cancela una solicitud mientras está pendiente; queda en estado terminal `cancelled` y el cuidador deja de verla como pendiente.",
        "operationId": "MarketplaceController_cancel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-09 A2 · Cancelar solicitud pendiente (solo el solicitante)",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/api/v1/hiring-requests/{id}/complete": {
      "post": {
        "operationId": "MarketplaceController_complete_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-09 · Finalizar / marcar como pagada (cierra el servicio)",
        "tags": [
          "Marketplace"
        ]
      }
    },
    "/api/v1/patients/{patientId}/caregivers": {
      "get": {
        "operationId": "MarketplaceController_history_v1",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaregiverHistoryItemDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-16 · Cuidadores del paciente (vigentes e históricos)",
        "tags": [
          "Marketplace"
        ]
      }
    },
...
  },
  "info": {
    "title": "Keru API",
    "description": "Backend del MVP de Keru (marketplace de cuidadores). Contrato para clientes (app móvil y web). Ver casos de uso en Keru-Casos-de-Uso-MVP.md y reglas en constitution.md.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "account": {
        "type": "apiKey",
        "in": "header",
        "name": "x-account-id"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CertificationDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Enfermería"
          },
          "institution": {
            "type": "string",
            "example": "Universidad de Buenos Aires"
          },
          "year": {
            "type": "number",
            "example": 2015
          }
        },
        "required": [
          "type",
          "institution",
          "year"
        ]
      },
      "AvailabilityDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "number",
            "example": 1,
            "description": "0=domingo .. 6=sábado"
          },
          "from": {
            "type": "string",
            "example": "08:00"
          },
          "to": {
            "type": "string",
            "example": "16:00"
          }
        },
        "required": [
          "dayOfWeek",
          "from",
          "to"
        ]
      },
      "RatesDto": {
        "type": "object",
        "properties": {
          "ratePerHour": {
            "type": "number",
            "example": 3500
          },
          "currency": {
            "type": "string",
            "example": "ARS",
            "default": "ARS"
          },
          "description": {
            "type": "string",
            "example": "Incluye acompañamiento nocturno"
          }
        },
        "required": [
          "ratePerHour"
        ]
      },
      "RegisterCaregiverDto": {
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "description": "Identidad de operación provista por el cliente (NFR-34). Un reintento con el mismo valor no duplica el efecto.",
            "example": "op-8f3a2c1e"
          },
          "displayName": {
            "type": "string",
            "example": "Laura Gómez"
          },
          "photoUrl": {
            "type": "string",
            "example": "http://localhost:4566/keru-media/images/abc.jpg"
          },
          "specialties": {
            "type": "array",
            "example": [
              "elder-care",
              "palliative"
            ],
            "items": {
              "type": "string",
              "enum": [
                "elder-care",
                "post-surgical",
                "chronic-illness",
                "disability",
                "palliative",
                "pediatric",
                "rehabilitation",
                "companionship"
              ]
            }
          },
          "certifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CertificationDto"
            }
          },
          "availability": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AvailabilityDto"
            }
          },
          "rates": {
            "$ref": "#/components/schemas/RatesDto"
          },
          "zone": {
            "type": "string",
            "example": "Palermo, CABA"
          },
          "modalities": {
            "type": "array",
            "example": [
              "home",
              "hospital"
            ],
            "items": {
              "type": "string",
              "enum": [
                "home",
                "hospital"
              ]
            }
          }
        },
        "required": [
          "operationId",
          "displayName",
          "specialties",
          "certifications",
          "availability",
          "rates",
          "zone",
          "modalities"
        ]
      },
      "CaregiverResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "deactivated"
            ]
          },
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "zone": {
            "type": "string"
          },
          "modalities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "badges": {
            "type": "object",
            "example": {
              "certifications": true,
              "identity": false,
              "background": false
            }
          },
          "rejectionReason": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "displayName",
          "status",
          "specialties",
          "zone",
          "modalities",
          "badges"
        ]
      },
...
      "CaregiverCardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string",
            "description": "Foto de perfil (UC-06)"
          },
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "zone": {
            "type": "string"
          },
          "modalities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ratePerHour": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "badges": {
            "type": "object"
          },
          "isFavorite": {
            "type": "boolean",
            "description": "true si está en los favoritos de la cuenta"
          },
          "ratingAverage": {
            "type": "number",
            "description": "Promedio de reseñas reveladas, 0 si no tiene (UC-06 criterio 3)"
          },
          "ratingCount": {
            "type": "number",
            "description": "Cantidad de reseñas reveladas"
          }
        },
        "required": [
          "id",
          "displayName",
          "specialties",
          "zone",
          "modalities",
          "ratePerHour",
          "currency",
          "badges"
        ]
      },
      "CaregiverProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string",
            "description": "Foto de perfil (UC-06)"
          },
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "zone": {
            "type": "string"
          },
          "modalities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ratePerHour": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "badges": {
            "type": "object"
          },
          "isFavorite": {
            "type": "boolean",
            "description": "true si está en los favoritos de la cuenta"
          },
          "ratingAverage": {
            "type": "number",
            "description": "Promedio de reseñas reveladas, 0 si no tiene (UC-06 criterio 3)"
          },
          "ratingCount": {
            "type": "number",
            "description": "Cantidad de reseñas reveladas"
          },
          "certifications": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "availability": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "id",
          "displayName",
          "specialties",
          "zone",
          "modalities",
          "ratePerHour",
          "currency",
          "badges",
          "certifications",
          "availability"
        ]
      },
      "CreateRequestDto": {
        "type": "object",
        "properties": {
          "operationId": {
            "type": "string",
            "description": "Identidad de operación provista por el cliente (NFR-34). Un reintento con el mismo valor no duplica el efecto.",
            "example": "op-8f3a2c1e"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "caregiverId": {
            "type": "string",
            "format": "uuid"
          },
          "modality": {
            "type": "string",
            "enum": [
              "home",
              "hospital"
            ]
          },
          "startDate": {
            "type": "string",
            "example": "2026-08-01T08:00:00Z"
          },
          "endDate": {
            "type": "string",
            "example": "2026-08-15T18:00:00Z"
          },
          "specialRequirements": {
            "type": "string",
            "example": "Requiere movilidad reducida"
          },
          "contactData": {
            "type": "object",
            "example": {
              "phone": "+54 11 5555-5555"
            }
          }
        },
        "required": [
          "operationId",
          "patientId",
          "caregiverId",
          "modality",
          "startDate",
          "endDate",
          "contactData"
        ]
      },
      "RequestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "caregiverId": {
            "type": "string",
            "format": "uuid"
          },
          "patientName": {
            "type": "string",
            "description": "Nombre del paciente (visible para el cuidador, UC-10)"
          },
          "caregiverName": {
            "type": "string",
            "description": "Nombre del cuidador (visible para el solicitante)"
          },
          "modality": {
            "type": "string"
          },
          "startDate": {
            "format": "date-time",
            "type": "string"
          },
          "endDate": {
            "format": "date-time",
            "type": "string"
          },
          "specialRequirements": {
            "type": "string"
          },
          "contactData": {
            "type": "object",
            "description": "Datos de contacto del solicitante. Para el cuidador solo con solicitud aceptada/en curso (UC-10)."
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "declined",
              "cancelled",
              "in-progress",
              "finished",
              "expired"
            ]
          },
          "ratePerHourSnapshot": {
            "type": "string",
            "description": "Tarifa pinneada al solicitar (NFR-03/23)"
          }
        },
        "required": [
          "id",
          "patientId",
          "caregiverId",
          "modality",
          "startDate",
          "endDate",
          "status",
          "ratePerHourSnapshot"
        ]
      },
      "CaregiverHistoryItemDto": {
        "type": "object",
        "properties": {
          "assignmentId": {
            "type": "string",
            "format": "uuid"
          },
          "caregiverId": {
            "type": "string",
            "format": "uuid"
          },
          "caregiverName": {
            "type": "string"
          },
          "periodStart": {
            "format": "date-time",
            "type": "string"
          },
          "periodEnd": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "historical"
            ]
          }
        },
        "required": [
          "assignmentId",
          "caregiverId",
          "periodStart",
          "periodEnd",
          "status"
        ]
      },
...
    }
  }
}